+++ jib-core/src/test/java/com/google/cloud/tools/jib/builder/steps/BuildImageStepTest.java
@@ -17,8 +17,9 @@
 package com.google.cloud.tools.jib.builder.steps;
 
 import com.google.cloud.tools.jib.blob.Blob;
+import com.google.cloud.tools.jib.blob.BlobDescriptor;
 import com.google.cloud.tools.jib.blob.Blobs;
-import com.google.cloud.tools.jib.cache.CacheEntry;
+import com.google.cloud.tools.jib.cache.CachedLayer;
 import com.google.cloud.tools.jib.configuration.BuildConfiguration;
 import com.google.cloud.tools.jib.configuration.ContainerConfiguration;
 import com.google.cloud.tools.jib.event.EventDispatcher;
@@ -63,26 +64,31 @@ public class BuildImageStepTest {
     testDescriptorDigest =
         DescriptorDigest.fromHash(
             "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa");
-    CacheEntry testCacheEntry =
-        new CacheEntry() {
+    CachedLayer testCachedLayer =
+        new CachedLayer() {
           @Override
-          public DescriptorDigest getLayerDigest() {
+          public DescriptorDigest getDigest() {
             return testDescriptorDigest;
           }
 
           @Override
-          public DescriptorDigest getLayerDiffId() {
+          public DescriptorDigest getDiffId() {
             return testDescriptorDigest;
           }
 
           @Override
-          public long getLayerSize() {
+          public long getSize() {
             return 0;
           }
 
           @Override
-          public Blob getLayerBlob() {
+          public Blob getBlob() {
             return Blobs.from("ignored");
+          }
+
+          @Override
+          public BlobDescriptor getBlobDescriptor() {
+            return new BlobDescriptor(0, testDescriptorDigest);
           }
         };
 
@@ -123,7 +129,7 @@ public class BuildImageStepTest {
             .addHistory(emptyLayerHistory)
             .build();
     Mockito.when(mockPullAndCacheBaseImageLayerStep.getFuture())
-        .thenReturn(Futures.immediateFuture(testCacheEntry));
+        .thenReturn(Futures.immediateFuture(testCachedLayer));
     Mockito.when(mockPullAndCacheBaseImageLayersStep.getFuture())
         .thenReturn(
             Futures.immediateFuture(
@@ -136,7 +142,7 @@ public class BuildImageStepTest {
             Futures.immediateFuture(
                 new PullBaseImageStep.BaseImageWithAuthorization(baseImage, null)));
     Mockito.when(mockBuildAndCacheApplicationLayerStep.getFuture())
-        .thenReturn(Futures.immediateFuture(testCacheEntry));
+        .thenReturn(Futures.immediateFuture(testCachedLayer));
   }
 
   @Test